requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
rich>=13.7.0
click>=8.1.0
python-dateutil>=2.9.0
